requirements.txt
selectolax>=0.3.0
hrequests>=0.1.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
email-validator>=2.0.0